Refactor base manager
